win-capture: deal with a 32bit handle cast errors
The HWND type is a void pointer, but HWND values are global and always
32bit despite, so casting to 32bit can cause cast warnings on actual
good compilers like gcc via mingw.  This change correctly handles the
casting to 32bits without producing unwanted warnings or errors on
mingw.

+++ plugins/win-capture/graphics-hook/d3d9-capture.cpp
@@ -285,7 +285,7 @@ static bool d3d9_shtex_init(uint32_t cx, uint32_t cy, HWND window)
 	}
 	if (!capture_init_shtex(&data.shtex_info, window, cx, cy,
 				data.cx, data.cy, data.dxgi_format, false,
-				(uint32_t)data.handle)) {
+				(uintptr_t)data.handle)) {
 		return false;
 	}
